+++ bc_obps/registration/tests/endpoints/v1/test_operators_endpoints.py
@@ -2,7 +2,7 @@ from registration.enums.enums import AccessRequestStates, AccessRequestTypes
 from model_bakery import baker
 from localflavor.ca.models import CAPostalCodeField
 from registration.tests.utils.bakers import operator_baker, user_operator_baker
-from registration.constants import AUDIT_FIELDS
+from common.constants import AUDIT_FIELDS
 from registration.models import Operator, UserOperator
 from registration.schema.v1.operator import OperatorOut, OperatorSearchOut, ConfirmSelectedOperatorOut
 from registration.tests.utils.helpers import CommonTestSetup, TestUtils
